build script improvement

+++ spark-doris-connector/build.sh
@@ -190,21 +190,26 @@ echo_g " scala version: ${SCALA_VERSION}"
 echo_g " spark version: ${SPARK_VERSION}, minor version: ${SPARK_MINOR_VERSION}"
 echo_g " build starting..."
 
-${MVN_BIN} clean package \
-  -Dspark.version=${SPARK_VERSION} \
-  -Dscala.version=${SCALA_VERSION} \
-  -Dthrift.binary=${THRIFT_BIN} \
-  -Dspark.minor.version=${SPARK_MINOR_VERSION} "$@"
+EXIT_CODE=$?
+if [ $EXIT_CODE -eq 0 ]; then
+  ${MVN_BIN} clean package \
+    -Dspark.version=${SPARK_VERSION} \
+    -Dscala.version=${SCALA_VERSION} \
+    -Dthrift.binary=${THRIFT_BIN} \
+    -Dspark.minor.version=${SPARK_MINOR_VERSION} "$@"
 
-DIST_DIR=${DORIS_HOME}/dist
-[ ! -d "$DIST_DIR" ] && mkdir "$DIST_DIR"
-dist_jar=$(ls "${ROOT}"/target | grep "spark-doris-" | grep -v "sources.jar" | grep -v "original-")
-rm -rf "${DIST_DIR}"/"${dist_jar}"
-cp "${ROOT}"/target/"${dist_jar}" "$DIST_DIR"
+  DIST_DIR=${DORIS_HOME}/dist
+  [ ! -d "$DIST_DIR" ] && mkdir "$DIST_DIR"
+  dist_jar=$(ls "${ROOT}"/target | grep "spark-doris-" | grep -v "sources.jar" | grep -v "original-")
+  rm -rf "${DIST_DIR}"/"${dist_jar}"
+  cp "${ROOT}"/target/"${dist_jar}" "$DIST_DIR"
 
-echo_g "*****************************************************************"
-echo_g "Successfully build Spark-Doris-Connector"
-echo_g "dist: $DIST_DIR/$dist_jar "
-echo_g "*****************************************************************"
-
-exit 0
+  echo_g "*****************************************************************"
+  echo_g "Successfully build Spark-Doris-Connector"
+  echo_g "dist: $DIST_DIR/$dist_jar "
+  echo_g "*****************************************************************"
+  exit 0;
+else
+  echo_r "Failed build Spark-Doris-Connector"
+  exit $EXIT_CODE;
+fi
